test: make latency test deterministic (#5895)
Description
---
Updates a latency test to be deterministic and exercise more cases.

Motivation and Context
---
A recent update in #5890 adds a test for sorting peers by latency. Its
test is randomized and may not fully exercise all the desired cases.
This PR makes a simple change that updates the test to be deterministic.

How Has This Been Tested?
---
It's a test, which passes!

What process can a PR reviewer use to test or verify this change?
---
Confirm that the test does what it says it does.