Change transaction protocol to bytes
Changed the transaction protocol in the wallet database to use bytes instead of
a hex string, as the underlying data type is encrypted bytes. This issue was
highlighted due to the `to_hex` function in `tari_utilities` not being able
to convert large transactions into hex strings (`**String to large**`) during
system-level coin-split stress testing.