feat!: completed transaction use bytes for transaction protocol (not hex string) in wallet database (#5906)
Description
---
Changed the transaction protocol in the wallet database to use bytes
instead of a hex string, as the underlying data type is encrypted bytes.
This issue was highlighted due to the `to_hex` function in
`tari_utilities` not being able to convert large transactions into hex
strings (returned `**String to large**`) for saving in the wallet
database during system-level coin-split stress testing.

Motivation and Context
---
See above.

How Has This Been Tested?
---
Existing unit tests and cucumber tests passed
Added two new unit tests and a cucumber test (long-running)
System-level test creating a coin-split transaction with 499 outputs
passed

What process can a PR reviewer use to test or verify this change?
---
Code walk-through
Run the new unit tests
Perform a system-level test creating a coin-split transaction with 499
outputs and wait for it to be mined and confirmed,

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [X] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
BREAKING CHANGE: All completed transactions in the wallet database will
be deleted at startup, while subsequent output validation will
re-allocate outputs that were encumbered to be spent as available and
invalidate outputs that were encumbered to be received.