feat!: add validator mmr size (#5873)
Description
---
Adds in the size of the validator balance tree to the header. 

Motivation and Context
---
The validator size needs to be committed too like the other Merkle roots
as well. A peer syncing will have no way of verifying the validator
merkle root is it does not know the size.

How Has This Been Tested?
---
Unit tests

What process can a PR reviewer use to test or verify this change?
---
ensure its committed to and not malleable.