feat: sender and receiver protocols use bytes (not hex string) in wallet database (#5950)
Description
---
Changed the sender and receiver protocol in the wallet database to use
bytes instead of a hex string, as the underlying data type is encrypted
bytes. This issue was highlighted due to the to_hex function in
tari_utilities not being able to convert large transactions into hex
strings (returned **String to large**) for saving in the wallet
database.

Motivation and Context
---
See above.

How Has This Been Tested?
---
Existing unit tests and cucumber tests passed
Added a new integration-level unit test `async fn
large_interactive_transaction()` to test the conversion from pending
outgoing and incoming transactions to and from the database.

What process can a PR reviewer use to test or verify this change?
---
Code walk-through
Run the new unit test

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

Co-authored-by: SW van Heerden <swvheerden@gmail.com>