chore: simplify metrics-only blocks (#5952)
Description
---
Does minor refactoring of metrics-only blocks to make the analyzer
happy.

Motivation and Context
---
Recent work in #5944 makes metrics a non-default feature. This made the
analyzer unhappy due to some unused variables, but it turns out this can
be fixed while simplifying some of the handling of metrics-only blocks.
This PR makes these changes and keeps the analyzer happy.

How Has This Been Tested?
---
Existing tests pass.

What process can a PR reviewer use to test or verify this change?
---
Check that the new metrics-only block handling is identical to that
introduced by #5944.