test: add gRPC interceptor authentication tests (#5958)
Description
---
Adds gRPC authentication tests using the client and server interceptors.

Closes #5957.

Motivation and Context
---
It was identified in #5957 that integration tests do not exercise
`Basic` gRPC authentication, and that the client and server interceptors
have no authentication unit tests. Adding full integration tests would
be unnecessarily complex, so this PR adds unit tests directly to the
interceptors.

Note that there is an authentication mismatch that is allowed and
tested. This occurs when the client is configured with `Basic`
authentication and the server is configured with no authentication. It
seems fine to let this use case succeed.

How Has This Been Tested?
---
These are new tests, which all pass.

What process can a PR reviewer use to test or verify this change?
---
Check that the new tests properly exercise all relevant success and
failure modes.