This check is not needed with ordered keys/sigs
Remove the check that a key has been validated previously. With ordered
sigs and keys a key will only ever be checked once so we don't need to
confirm if it's already provided a valid sig.

It's also confusing because Alice can provide two sigs against her pub
key but she has to provide the pubkey twice. The language here made that
a litle confusing.