chore: migrate from `lazy_static` to `once_cell` (#5992)
Description
---
Migrates all remaining uses of `lazy_static` to `once_cell` for better
consistency with the rest of the ecosystem.

Motivation and Context
---
Other repositories in the ecosystem are moving from `lazy_static` to
`once_cell` for better maintenance and to remove the use of a macro.
This repository had a mix of both still in use, so this PR finishes
moving everything over. The API is very similar, so changes were
minimal.

How Has This Been Tested?
---
Existing tests pass.

What process can a PR reviewer use to test or verify this change?
---
Check the correctness of each change by asserting that each use of
`Lazy` reflects the same functionality as the previous `lazy_static!`
macro.