feat!: dont store entire monero coinbase transaction (#5991)
Description
---
only store the keccak state and not the entire monero coinbase
transaction

Motivation and Context
---
The monero coinbase transaction can be very large as it may contain
multiple coinbases. We only store the keccak inner hash state with half
of the hashed transaction. We hash the last part to verify its working.

How Has This Been Tested?
---
unit tests

fixes: #5891