feat: fix windows installer (#6043)
Description
---
A recent PR changed how the default `config.toml` file is created based
on user input, so that it cannot be created ahead of time anymore. This
PR applies that behaviour to the Windows installer.

Motivation and Context
---
See above

How Has This Been Tested?
---
Generated the install file, tested the installation as well as the
runtime

What process can a PR reviewer use to test or verify this change?
---
As above

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->