feat: make libtor on by default for nix builds (#6060)
Description
---
This changes the defaults to include libtor for the MinoTari Node and
Console wallet by default. Removing the need to run tor independently.

Closes #6050 

Motivation and Context
---
Make setup a smidge easier

How Has This Been Tested?
---
Locally: Ran the node and wallet without having a tor session started on
the machine. Node and wallet used the built in libtor no problem.

Changed `use_libtor` to `false` in the config. Started the node and
received a `ExitError { exit_code: TorOffline` error as expected.
Turned on local tor session, and restarted node with no issue. Node is
now running on provided tor, and not built in.

Windows built without error in CI:
https://github.com/brianp/tari/actions/runs/7274136545

@Cifko did the windows testing for me. He could compile natively as
usual without any libtor related problems.
He also ran the installed package from the test builds with no issue.

This confirms that us setting a libtor default for *nix systems is not
currently interfering with windows compilation or builds.

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify