chore(ci): fix windows s3 uploads error again (#6065)
Description
Target Windows installer for S3 upload.
Fix the scheduled target for daily uploads.
Updated the env for the S3 Destination Override to match other envs.

Motivation and Context
Fix the S3 upload for Windows.

How Has This Been Tested?
Tested in a branch for the repo.

What process can a PR reviewer use to test or verify this change?
Check the Workflow for any errors.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->