chore: upgrade tonic and prost (#6067)
Description
---
We were hoping to get moved onto a more modern version of both but this
will not be possible. Between versions 0.8.4 and 0.9.0 of tonic they
dropped support for the rust 2018 version. We would also need to upgrade
to 2021 to make further use of more recent versions of tonic.

This PR upgrades them to the most recent possible.

I additionally tried seeing if we could run older versions only _where
nedded_ and run more modern version in other places but most of our
tonic use is tied to the grpc and the grpc clients require the older
version, in turn most everything requires the old version.

Motivation and Context
---
Upgrade the old to something newer.

How Has This Been Tested?
---
CI

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

---------

Co-authored-by: SW van Heerden <swvheerden@gmail.com>