fix: wallet coinbases not validated correctly (#6074)
Description
---
Fixes adding inputs scanned from the block to be improperly marked on
adding to the database.

Motivation and Context
---
Inputs scanned, such as coinbases are marked as `unspent` while this is
technically true from the wallet's perspective as they are unspent and
are detected from a mined block. We need to mark them as
`unspentUnconfiremd`. We need to make sure they are properly confirmed
by the validation task. The validation task will pick up these outputs
and flag them as spent/reorged etc.

But in the current case because they are flagged as `unspent` the wallet
can select them when spending, although they might not have passed the
confirmation time, or the might have been reorged out of the main chain.