chore: add script to manage crates.io owners
There are many packages in the tari codebase. To prevent squatting and
possible domain hijacking attacks, we need to register these on
crates.io (they don't have namespacing :( )

Managing the owners of these crates on crates.io is a manual nightmare,
so this script can automate it for you.

This commit introduces significant refinements to the owner management script
update_owners.sh.

The script now iterates over packages, lists current owners, and compares them
with expected owners.

It includes functions to add, remove, and verify owners based on the comparison.
Packages with the group 'ignore' are skipped during the iteration.