chore: make MAC equality check more idiomatic (#6123)
Description
---
Cleans up the handling of cipher seed message authentication codes
(MACs) to be more idiomatic.

Motivation and Context
---
A cipher seed has an associated MAC that is checked during decryption in
constant time. Rather than using `Choice::unwrap_u8` as is currently
done, this PR uses the more idiomatic `Into<bool>` on `Choice` to be
more resistant to mistakes or misuse. It also changes the logic slightly
to avoid "failing open" by no longer returning an `Ok` at the end of the
function by default, which helps to simplify the code and make the
intent more clear.

How Has This Been Tested?
---
Existing tests pass.

What process can a PR reviewer use to test or verify this change?
---
Check that the updated logic matches the existing logic.