feat: implement inflating tail emission
This PR primarily adds tail emission inflation as a featuer; but it does
so using feature gates, which were broken in a few places, and so a big
part of this PR is resolving issues related to feature gates.

== Tail emission inflation

Adds a feature tari_feature_mainnet_emission to allow for tail emission
inflation. See https://github.com/tari-project/tari/issues/6122

This change necessitates the addition of 2 new consensus constants:
`inflation_bips` -- the annual inflation rate of the total supply.
and `tail_emission_epoch_length`, which controls the tail emission
inflation. These replace `tail_emission`.

We update the Protobuf definition for ConsensusConstants to account for
the new fields.

== Getting Feature flags working

When formally implementing compiler features, the setting of the `TARI_NETWORK`
and (from now) `TARI_TARGET_NETWORK` envar and the related compiler flags matter.

These changes require that when tests run, the correct network flags are set.

For example, blocks::genesis_block::test::stagenet_genesis_sanity_check should
only run when TARI_NETWORK=stagenet.

While writing this PR, I uncovered a problem:
There is not a 1:1 mapping between the build _target_ and the actual
network the binary may run on. E.g. the mainnet build can run on either
stagenet or mainnet.

Unfortunately the `TARI_NETWORK` envar was used to set both of these
variables. Without using feature flags, this discrepency went unnoticed.
With feature flags, a lot of tests that implicitly assumed say a mainnet
target and a stagenet run would simply fail, with no way to resolve the
issue.

Since these are 2 different things, this commit introduced
`TARI_TARGET_NETWORK` whicg sets the **build target**. In most contexts,
this then decouples the configured network, which is specified by the
`TARI_NETWORK` envar  as usual.

The other changes in this commit revolve around updating merkle roots
and the correct faucet sets to make the target/network/faucet
combination consistent.

== CI workflows

The CI is configured to run tests for each of
* Testnet (Esme)
* Nextnet (Nextnet)
* Mainnet (Mainnet, with some tests specifying stagenet)

=== remove env_logger

I used env_logger to try print the feature list at compiler time, but it
doesn't work, so rather remove it.

A better strategy (implemented in this commit) is to print
"tari_feature" as a prefix to the message emitted by build.rs. Then you
can do soemthing like

`cargo build -vv .... 2>/dev/null` | grep "tari_feature"`

Piping stderr to dev null hides all the noise from rustc, and then you
can grep for the messages you want (or omit the grep to see the usual
build messages among the rest)

===  set network according to pols

Sets the network according to the principle of least surprise. i.e. if
TARI_NETWORK is set, try and make that the default network.